pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wa-gov-au-harvest-consultations"
version = "0.3.1"
description = "Batch ingestion pipeline for harvesting consultation data from REST APIs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.5",
    "ibis-framework[duckdb]>=9.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["harvest", "models"]
